refactor(holiday): consume shared docx_report framework; drop local docx modules
