+++ tests/utils/holiday_report/test_fragments.py
@@ -1,4 +1,4 @@
-from epic_news.utils.holiday_report.fragments import generate_fragment
+from epic_news.utils.docx_report import generate_fragment
 
 
 class OkLLM:
@@ -17,15 +17,21 @@ class EmptyLLM:
 
 
 def test_fragment_returns_markdown_on_success():
-    md = generate_fragment("Restauration", "Liste les restaurants.", "recherche", OkLLM())
+    md = generate_fragment(
+        "Restauration", "Liste les restaurants.", "recherche", OkLLM(), system="Tu es un rédacteur de voyage."
+    )
     assert "Markdown" in md
 
 
 def test_fragment_returns_placeholder_on_failure():
-    md = generate_fragment("Budget", "Détaille le budget.", "recherche", FailLLM())
+    md = generate_fragment(
+        "Budget", "Détaille le budget.", "recherche", FailLLM(), system="Tu es un rédacteur de voyage."
+    )
     assert "indisponible" in md.lower()
 
 
 def test_fragment_returns_placeholder_on_empty():
-    md = generate_fragment("Budget", "Détaille le budget.", "recherche", EmptyLLM())
+    md = generate_fragment(
+        "Budget", "Détaille le budget.", "recherche", EmptyLLM(), system="Tu es un rédacteur de voyage."
+    )
     assert "indisponible" in md.lower()
